+++ resources/views/pages/produk/index.blade.php
@@ -84,7 +84,7 @@
                         let edit =
                             `<a title="Edit Data" href="/produk/${id}/edit" class="btn btn-warning me-2"><i class="fa fa-pencil"></i></a>`;
                         let deletebtn =
-                            `<a title="Hapus Data" href="/produk/delete/${id}" class="btn btn-danger"><i class="fa fa-trash"></i></a>`
+                            `<a onclick="return confirm('Data ini akan dihapus')" title="Hapus Data" href="/produk/delete/${id}" class="btn btn-danger"><i class="fa fa-trash"></i></a>`
                         return edit + deletebtn
                     },
                 },
